# openapi.yaml
openapi: 3.1.0
info:
  title: Manda Money API
  version: 1.0.0
  description: Versioned API contract replacing the Google Apps Script endpoints.
servers:
  - url: https://api.example.invalid/api/v1
    description: Replace with the deployed API origin.
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      security: []
      summary: Sign in with the existing phone-and-birthday credentials
      description: Verifies the legacy phone-and-birthday pair using its keyed verifier and returns an opaque revocable session. Login is rate limited.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, birthday]
              properties:
                phone:
                  type: string
                  minLength: 10
                  maxLength: 20
                  description: Phone number normalized to digits by the server.
                birthday:
                  type: string
                  pattern: '^\d{2}/\d{2}/\d{4}$'
      responses:
        "200":
          description: Authenticated user and opaque access token.
          content:
            application/json:
              schema:
                type: object
                required: [accessToken, expiresAt, user]
                properties:
                  accessToken: { type: string }
                  expiresAt: { type: string, format: date-time }
                  user: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /auth/me:
    get:
      summary: Return the authenticated user's identity
      responses:
        "200":
          description: Authenticated user identity.
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/logout:
    post:
      summary: Revoke the current API session
      responses:
        "200": { description: Session revoked. }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
...
  schemas:
...
    User:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    RateLimited:
      description: Too many requests.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...